requirements.txt
Django~=4.2.4
djangorestframework~=3.14.0
django-mptt~=0.14.0
python-decouple~=3.8
